tools/notion_sync/requirements.txt
requests
pyyaml
python-slugify
